+++ packages/python/plotly/plotly/express/_core.py
@@ -660,6 +660,12 @@ def configure_cartesian_axes(args, fig, orders):
     if "is_timeline" in args:
         fig.update_xaxes(type="date")
 
+    if "ecdfmode" in args:
+        if args["orientation"] == "v":
+            fig.update_yaxes(rangemode="tozero")
+        else:
+            fig.update_xaxes(rangemode="tozero")
+
 
 def configure_ternary_axes(args, fig, orders):
     fig.update_ternaries(
@@ -1312,6 +1318,7 @@ def build_dataframe(args, constructor):
     wide_cross_name = None  # will likely be "index" in wide_mode
     value_name = None  # will likely be "value" in wide_mode
     hist2d_types = [go.Histogram2d, go.Histogram2dContour]
+    hist1d_orientation = constructor == go.Histogram or "ecdfmode" in args
     if constructor in cartesians:
         if wide_x and wide_y:
             raise ValueError(
@@ -1346,7 +1353,7 @@ def build_dataframe(args, constructor):
                 df_provided and var_name in df_input
             ):
                 var_name = "variable"
-            if constructor == go.Histogram:
+            if hist1d_orientation:
                 wide_orientation = "v" if wide_x else "h"
             else:
                 wide_orientation = "v" if wide_y else "h"
@@ -1360,7 +1367,10 @@ def build_dataframe(args, constructor):
         var_name = _escape_col_name(df_input, var_name, [])
 
     missing_bar_dim = None
-    if constructor in [go.Scatter, go.Bar, go.Funnel] + hist2d_types:
+    if (
+        constructor in [go.Scatter, go.Bar, go.Funnel] + hist2d_types
+        and not hist1d_orientation
+    ):
         if not wide_mode and (no_x != no_y):
             for ax in ["x", "y"]:
                 if args.get(ax) is None:
@@ -1457,14 +1467,18 @@ def build_dataframe(args, constructor):
         df_output[var_name] = df_output[var_name].astype(str)
         orient_v = wide_orientation == "v"
 
-        if constructor in [go.Scatter, go.Funnel] + hist2d_types:
+        if hist1d_orientation:
+            args["x" if orient_v else "y"] = value_name
+            args["y" if orient_v else "x"] = wide_cross_name
+            args["color"] = args["color"] or var_name
+        elif constructor in [go.Scatter, go.Funnel] + hist2d_types:
             args["x" if orient_v else "y"] = wide_cross_name
             args["y" if orient_v else "x"] = value_name
             if constructor != go.Histogram2d:
                 args["color"] = args["color"] or var_name
             if "line_group" in args:
                 args["line_group"] = args["line_group"] or var_name
-        if constructor == go.Bar:
+        elif constructor == go.Bar:
             if _is_continuous(df_output, value_name):
                 args["x" if orient_v else "y"] = wide_cross_name
                 args["y" if orient_v else "x"] = value_name
@@ -1474,13 +1488,24 @@ def build_dataframe(args, constructor):
                 args["y" if orient_v else "x"] = count_name
                 df_output[count_name] = 1
                 args["color"] = args["color"] or var_name
-        if constructor in [go.Violin, go.Box]:
+        elif constructor in [go.Violin, go.Box]:
             args["x" if orient_v else "y"] = wide_cross_name or var_name
             args["y" if orient_v else "x"] = value_name
-        if constructor == go.Histogram:
-            args["x" if orient_v else "y"] = value_name
-            args["y" if orient_v else "x"] = wide_cross_name
-            args["color"] = args["color"] or var_name
+
+    if hist1d_orientation and constructor == go.Scatter:
+        if args["x"] is not None and args["y"] is not None:
+            args["histfunc"] = "sum"
+        elif args["x"] is None:
+            args["histfunc"] = None
+            args["orientation"] = "h"
+            args["x"] = count_name
+            df_output[count_name] = 1
+        else:
+            args["histfunc"] = None
+            args["orientation"] = "v"
+            args["y"] = count_name
+            df_output[count_name] = 1
+
     if no_color:
         args["color"] = None
     args["data_frame"] = df_output
@@ -1778,12 +1803,18 @@ def infer_config(args, constructor, trace_patch, layout_patch):
             trace_patch["opacity"] = args["opacity"]
         else:
             trace_patch["marker"] = dict(opacity=args["opacity"])
-    if "line_group" in args:  # px.line, px.line_*, px.area
-        modes = set(["lines"])
+    if (
+        "line_group" in args or "line_dash" in args
+    ):  # px.line, px.line_*, px.area, px.ecdf
+        modes = set()
+        if args.get("lines", True):
+            modes.add("lines")
         if args.get("text") or args.get("symbol") or args.get("markers"):
             modes.add("markers")
         if args.get("text"):
             modes.add("text")
+        if len(modes) == 0:
+            modes.add("lines")
         trace_patch["mode"] = "+".join(modes)
     elif constructor != go.Splom and (
         "symbol" in args or constructor == go.Scattermapbox
@@ -1792,6 +1823,10 @@ def infer_config(args, constructor, trace_patch, layout_patch):
 
     if "line_shape" in args:
         trace_patch["line"] = dict(shape=args["line_shape"])
+    elif "ecdfmode" in args:
+        trace_patch["line"] = dict(
+            shape="vh" if args["ecdfmode"] == "reversed" else "hv"
+        )
 
     if "geojson" in args:
         trace_patch["featureidkey"] = args["featureidkey"]
@@ -1832,6 +1867,14 @@ def infer_config(args, constructor, trace_patch, layout_patch):
 
     if "trendline_options" in args and args["trendline_options"] is None:
         args["trendline_options"] = dict()
+
+    if "ecdfnorm" in args:
+        if args.get("ecdfnorm", None) not in [None, "percent", "probability"]:
+            raise ValueError(
+                "`ecdfnorm` must be one of None, 'percent' or 'probability'. "
+                + "'%s' was provided." % args["ecdfnorm"]
+            )
+        args["histnorm"] = args["ecdfnorm"]
 
     # Compute applicable grouping attributes
     for k in group_attrables:
@@ -1984,11 +2027,11 @@ def make_figure(args, constructor, trace_patch=None, layout_patch=None):
                         trace_spec != trace_specs[0]
                         and (
                             trace_spec.constructor in [go.Violin, go.Box]
-                            and m.variable in ["symbol", "pattern"]
+                            and m.variable in ["symbol", "pattern", "dash"]
                         )
                         or (
                             trace_spec.constructor in [go.Histogram]
-                            and m.variable in ["symbol"]
+                            and m.variable in ["symbol", "dash"]
                         )
                     ):
                         pass
@@ -2046,6 +2089,24 @@ def make_figure(args, constructor, trace_patch=None, layout_patch=None):
                 and trace.line.color
             ):
                 trace.update(marker=dict(color=trace.line.color))
+
+            if "ecdfmode" in args:
+                base = args["x"] if args["orientation"] == "v" else args["y"]
+                var = args["x"] if args["orientation"] == "h" else args["y"]
+                ascending = args.get("ecdfmode", "standard") != "reversed"
+                group = group.sort_values(by=base, ascending=ascending)
+                group_sum = group[var].sum()  # compute here before next line mutates
+                group[var] = group[var].cumsum()
+                if not ascending:
+                    group = group.sort_values(by=base, ascending=True)
+
+                if args.get("ecdfmode", "standard") == "complementary":
+                    group[var] = group_sum - group[var]
+
+                if args["ecdfnorm"] == "probability":
+                    group[var] = group[var] / group_sum
+                elif args["ecdfnorm"] == "percent":
+                    group[var] = 100.0 * group[var] / group_sum
 
             patch, fit_results = make_trace_kwargs(
                 args, trace_spec, group, mapping_labels.copy(), sizeref